Remove ElementCallArgumentFactory
